fixed compiler warning

+++ mobile/ios/Classes/KeychainItemWrapper.m
@@ -306,7 +306,7 @@
         
         result = SecItemUpdate((CFDictionaryRef)updateItem, (CFDictionaryRef)tempCheck);
         if (result != noErr) {
-            NSLog(@"error updating keychain item: %ld", result);
+            NSLog(@"error updating keychain item: %d", (int) result);
             NSLog(@"updateItem: %@", [updateItem description]);
         }
     }
@@ -315,7 +315,7 @@
         // No previous item found; add the new one.
         result = SecItemAdd((CFDictionaryRef)[self dictionaryToSecItemFormat:keychainItemData], NULL);
         if (result != noErr) {
-            NSLog(@"error creating keychain item: %ld", result);
+            NSLog(@"error creating keychain item: %d", (int) result);
             NSLog(@"keychainItemData: %@", [keychainItemData description]);
         }
     }
